+++ transport/tchannel/handler_test.go
@@ -74,7 +74,6 @@ func TestHandlerErrors(t *testing.T) {
 		registry.EXPECT().GetHandler("service", "hello").Return(rpcHandler, nil)
 		rpcHandler.EXPECT().Handle(
 			transporttest.NewContextMatcher(t),
-			transportOptions,
 			transporttest.NewRequestMatcher(t,
 				&transport.Request{
 					Caller:    "caller",
@@ -191,7 +190,6 @@ func TestHandlerFailures(t *testing.T) {
 			expectCall: func(h *transporttest.MockHandler) {
 				h.EXPECT().Handle(
 					transporttest.NewContextMatcher(t, transporttest.ContextTTL(time.Second)),
-					transportOptions,
 					transporttest.NewRequestMatcher(
 						t, &transport.Request{
 							Caller:    "bar",
@@ -229,7 +227,6 @@ func TestHandlerFailures(t *testing.T) {
 				}
 				h.EXPECT().Handle(
 					transporttest.NewContextMatcher(t, transporttest.ContextTTL(time.Second)),
-					transportOptions,
 					transporttest.NewRequestMatcher(t, req),
 					gomock.Any(),
 				).Return(
@@ -269,15 +266,9 @@ func TestHandlerFailures(t *testing.T) {
 				h.EXPECT().Handle(
 					transporttest.NewContextMatcher(
 						t, transporttest.ContextTTL(time.Millisecond)),
-					transportOptions,
 					transporttest.NewRequestMatcher(t, req),
 					gomock.Any(),
-				).Do(func(
-					ctx context.Context,
-					_ transport.Options,
-					_ *transport.Request,
-					_ transport.ResponseWriter,
-				) {
+				).Do(func(ctx context.Context, _ *transport.Request, _ transport.ResponseWriter) {
 					<-ctx.Done()
 				}).Return(context.DeadlineExceeded)
 			},
@@ -306,15 +297,9 @@ func TestHandlerFailures(t *testing.T) {
 				h.EXPECT().Handle(
 					transporttest.NewContextMatcher(
 						t, transporttest.ContextTTL(time.Second)),
-					transportOptions,
 					transporttest.NewRequestMatcher(t, req),
 					gomock.Any(),
-				).Do(func(
-					_ context.Context,
-					_ transport.Options,
-					_ *transport.Request,
-					_ transport.ResponseWriter,
-				) {
+				).Do(func(context.Context, *transport.Request, transport.ResponseWriter) {
 					panic("oops I panicked!")
 				})
 			},