+++ transport/tchannel/outbound.go
@@ -87,10 +87,6 @@ func (o outbound) Stop() error {
 	return nil
 }
 
-func (outbound) Options() transport.Options {
-	return transportOptions
-}
-
 func (o outbound) Call(ctx context.Context, req *transport.Request) (*transport.Response, error) {
 	if !o.started.Load() {
 		// panic because there's no recovery from this