+++ CHANGELOG.md
@@ -23,6 +23,8 @@ This release requires regeneration of ThriftRW code.
     clients and servers.
 -   **Breaking**: Remove the ability to set Baggage Headers through yarpc, use
     opentracing baggage instead
+-   **Breaking**: Transport options have been removed completely. Encoding
+    values differently based on the transport is no longer supported.
 
 
 v0.3.1 (2016-09-31)